check outflow by avgprice instead of exchange price
